+++ src/pages/SamplePage.vue
@@ -1,5 +1,5 @@
 <script lang="ts" setup>
-import PickButton from "@/components/button/PickButton.vue";
+import PickButton from "@/components/button/ReviewPickButton.vue";
 import SubscribeButton from "@/components/button/SubscribeButton.vue";
 import ReviewRegisterButton from "@/components/button/ReviewRegisterButton.vue";
 import CalendarIcon from "@/components/calendar/CalendarIcon.vue";
