--- src/types/Review.ts
@@ -1,12 +1,0 @@
-export interface Review {
-  reviewId: number;
-  memberId: number;
-  nickname: string;
-  bookId: number;
-  bookTitle: string;
-  reviewTitle: string;
-  reviewContent: string;
-  pickCount: number;
-  createdAt: Date;
-  updatedAt: Date;
-}
